requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1  # Corrected package name and version
mysqlclient==2.1.1
pyotp==2.9.0
requests==2.31.0
django-redis==5.4.0
google-auth==2.23.0
python-decouple==3.8
gunicorn==22.0.0